+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 0.7.0
 
+- *BREAKING* `client.current_user` is not a method, not a property for more consistency.
 - Joining of ListResult objects with `__add__` to aid pagination of results, with example in README.
 
 ## 0.6.1